LayoutTests/http/tests/site-isolation/inspector/debugger/scriptParsed-frame-target-expected.txt

Test that a frame target receives Debugger.scriptParsed events and scripts are isolated from the main target.

== Running test suite: SiteIsolation.Debugger.ScriptParsed
-- Running test case: SiteIsolation.Debugger.ScriptParsed.FrameTargetCreated
PASS: Added target should have Frame type.
PASS: Cross-origin frame target should have a DebuggerAgent.

-- Running test case: SiteIsolation.Debugger.ScriptParsed.FrameTargetHasScripts
PASS: Frame target should have at least one parsed script.
PASS: Frame target should have the debugger-frame-script.js script.

-- Running test case: SiteIsolation.Debugger.ScriptParsed.ScriptsIsolatedFromMainTarget
PASS: Main target should not have the debugger-frame-script.js script.

-- Running test case: SiteIsolation.Debugger.ScriptParsed.RemoveIframe
PASS: Removed target should have Frame type.
